requirements.txt
streamlit
google-genai
python-dotenv
pandas
gspread
oauth2client